pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "task-scheduler-env"
version = "0.1.0"
description = "Sprint Task Scheduler OpenEnv environment"
requires-python = ">=3.10"

dependencies = [
    "openenv-core>=0.2.3",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
    "typing-extensions>=4.8.0",
    "websockets>=12.0",
    "platformdirs>=3.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["server"]
package-dir = {"server" = "server"}
